hopefully_final_commit-fixed_design_and_insert

+++ CLASSES/delete_query.php
@@ -1,16 +1,26 @@
 <?php
-
-require_once('Con_class.php');
-class querydelete_class extends Con_class
+if(isset($_POST['delete-checkbox[]']))
 {
-    public function delete_query($IDs)
+    require_once('Con_class.php');
+    class querydelete_class extends Con_class
     {
-            $ID = implode(",",$IDs);
-            $sqlquery = "DELETE FROM products WHERE prod_ID IN ($ID)";
-            $stament = $this -> concexecution()->prepare($sqlquery);
-            $stament -> execute();
-            header("Location: ../index.php");
+        public function delete_query($IDs)
+
+        {
+
+                $ID = implode(",",$IDs);
+                $sqlquery = "DELETE FROM products WHERE prod_ID IN ($ID)";
+                $stament = $this -> concexecution()->prepare($sqlquery);
+                $stament -> execute();
+                header("Location: ../index.php");
+        }
     }
+    $deleteing = new querydelete_class();
+    $deleteing->delete_query($_POST['delete-checkbox']);
+
 }
-$deleteing = new querydelete_class();
-$deleteing->delete_query($_POST['delete-checkbox']);
+
+else
+{
+    header("Location: ../index.php");    
+}
